feat: extension for opening link in Tauri
