+++ client/.env
@@ -1,2 +1,3 @@
+VITE_GITHUB_REPO_URL="https://www.github.com/jukrb0x/taskhorizon"
 VITE_BASE_URL="http://localhost:8083" # Backend API Base URL
 VITE_APP_VERSION="0.2.5" # App Version
