+++ client/src-tauri/tauri.conf.json
@@ -11,7 +11,10 @@
     },
     "tauri": {
         "allowlist": {
-            "all": true
+            "all": true,
+            "shell": {
+                "open": true
+            }
         },
         "bundle": {
             "active": true,
