+++ client/src/components/Settings/SettingsProfile.tsx
@@ -1,3 +1,3 @@
 export const SettingsProfile = () => {
-    return <> profile</>;
+    return <>aaa</>;
 };
