+++ client/src/hooks/use-tauri-extension.tsx
@@ -1,8 +1,10 @@
+import { open } from '@tauri-apps/api/shell';
 import { useEffect, useState } from 'react';
 
 const useTauriExtension = () => {
     const [isTauri, setIsTauri] = useState(false);
 
+    // detect tauri on runtime
     useEffect(() => {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
@@ -11,7 +13,11 @@ const useTauriExtension = () => {
         }
     }, []);
 
-    return isTauri;
+    const openLink = async (url: string, isNewTab = true) => {
+        isTauri ? await open(url) : window.open(url, isNewTab ? '_blank' : '_self');
+    };
+
+    return { isTauri, openLink };
 };
 
 export { useTauriExtension };
